pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teacher3d"
version = "0.1.0"
description = "Teacher-guided occlusion-aware single-image scene reconstruction scaffold"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "Pillow>=9.0",
  "PyYAML>=6.0",
  "torch>=2.6",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
